requirements.txt
pyyaml>=6.0
aiohttp>=3.8.0
packaging>=21.0
requests>=2.25.0
